Rename exercises to sendExerciseReadyForApproval
